DO NOT MERGE: disable connectivity-tiramisu build
Disable framework-connectivity-tiramisu and do not include it in the
apex, and add empty filegroups and replacement stub classes to make
service-connectivity-tiramisu only contain
ConnectivityServiceInitializer.

This allows continuing to build the com.android.tethering APEX in the S
release branch, while minimizing build conflicts as T classes are being
prepared to move to it.

Bug: 206893064
Test: m com.android.tethering
Ignore-AOSP-First: sc-mainline-prod only.
Change-Id: I0c4e29568eadb6ef85ddadebbb37723adb9f29cd