Add service.proto jarjar rule
This brings the jarjar rule in line with contents of downstream
branches.

The change should be a no-op, since there is no class in that package in
branches that do not have the jarjar rule.

(this is a partial cherry-pick of change:
I947c9e3b6ebc232e230782f6345c5cb6893138e6)

Test: m

Merged-In: I947c9e3b6ebc232e230782f6345c5cb6893138e6
Change-Id: I0d9f8944572f87e0bf52862ba5e4382386b161a3