Do not set service-nearby-pre-jarjar installable
service-nearby-pre-jarjar is just an build intermediate: the classes are
eventually added to service-connectivity.jar, which is what gets shipped
on device. So service-nearby-pre-jarjar should not be installable.

Marking it installable causes it to be dexed, which is inconsistent with
other service-connectivity intermediates. Having it be a plain jar
should be faster to build and make inspection / jarjar rule generation
easier.

Ignore-AOSP-First: Nearby not yet in AOSP
Bug: 217129444
Test: m
Change-Id: I2f6234d488edf25e600147ec8bf2961af9929cad