Explicitly enabled Proguard compat for NearbyMultiDevicesClientsSnippets
This target statically links MessageNano, colliding with the system
version of that class. Normally this is resolved through jarjar, but
as a test, and with shrinking/optimization disabled, this collision
should be benign. Enable R8 compat mode to treat the failure as a
warning.

Bug: 215530220
Change-Id: Id09d839dbe2ef7e2f98e6a79f1338078f1b24378
Test: m NearbyMultiDevicesClientsSnippets
Ignore-AOSP-First: nearby_not_in_aosp_yet