Clear DNS cache after every network access.
If the app tries to make a network connection in the
foreground immediately after trying to do the same when
it's network access was blocked, it could receive a
UnknownHostException due to the cached DNS entry.
So, clear the dns cache after every network access for
now until we have a fix on the platform side.

Bug: 237549250
Bug: 238064333
Bug: 237556596
Bug: 238962235
Bug: 238833668
Bug: 239262832
Test: atest ./tests/cts/hostside/src/com/android/cts/net/HostsideRestrictBackgroundNetworkTests.java
Ignore-AOSP-First: Will be cherrypicked to aosp later
Change-Id: Ic603b1bd9da24ee3724054c4a91794ee569e97f6