Run R8 with necessary Tethering classpath classes
R8 emits "missing class" warnings when optimizing Tethering.apk
because Tethering packages the already built TetheringApiXLib, but does
not specify any classpath library, which are actually needed for R8.

This causes wrong optimizations, where R8 would sometimes strip out code
that it thinks is unused (because it considers it unreachable due to the
missing class, or it does not know the superclass of a callback and
thinks the overridden methods cannot be called).

Add the necessary classpath libraries to remove the warnings.

Bug: 226127213
Test: m
Change-Id: Icd639d2783a2adc21e05a3b44f276ad72ad1e2d3