ethernet: reenable IPv6 provisioning delays
Unfortunately, there is no way to disable carrier on the tap interface
without first configuring the interface with carrier. This means that it
is possible the interface is brought up before carrier is disabled (it
is an unlikely flake but it does happen in automation). Reenabling RS
delay and DAD will make sure that provisioning does not succeed before
carrier is disabled.

Test: atest EthernetManagerTest --iterations 50
Change-Id: Ie2aa7529b48fa9aceff3dd3b432d87a6eb2f0298