Improve the dependency declaration for Net tests
This change moves ConnectivityChecker from `required` to `data`,
since the `required` is inherited from Make and is very vague. It's
discouraged by Android Build team to use it.

Besides, this change declares net-tests-utils-host-common, since Net
tests use ConnectivityCheckTargetPreparer that is included in
net-tests-utils-host-common. All dependencies of a test should be
declared explicitly, so that test infrastructure could only provide
the needed dependencies instead of all to run a test.

Bug: 222438068
Test: presubmit
Change-Id: Id522fd88ca53b5af338315fc6370d68bdabf403b