gn2bp: add basic action support
WARNING: this is completely untested and probably wrong in more ways
than one. Actions are not universally convertable to soong, so
"additional_args" need to be configured.

This does get us one step closer to running the script without any
errors. No new targets are supported with this CL though.

Test: //base:base
Change-Id: I38a0e05f5e9b2153fc667882f4b35de3e330f898